Add files via upload
The 'rockpaperscissors' dataset is popularly used in machine learning for image recognition and classification. This dataset contains images of hands forming the symbols of rock, paper, and scissors, which are typically used in the rock-paper-scissors game. Each image is labeled to indicate the action performed (e.g., rock, paper, or scissors).